+++ src/pysampling/sample.py
@@ -1,6 +1,7 @@
 from pysampling.algorithms.halton import HaltonSampling
 from pysampling.algorithms.lhs import LatinHypercubeSampling
 from pysampling.algorithms.random import RandomSampling
+from pysampling.algorithms.riesz import RieszEnergySampling
 from pysampling.algorithms.sobol import SobolSampling
 from pysampling.sampling import Sampling
 
@@ -20,10 +21,13 @@ def sample(algorithm, n_points, n_dim, *args, **kwargs):
 
     elif algorithm == "sobol":
         _obj = SobolSampling(*args, **kwargs)
+
+    elif algorithm == "riesz":
+        _obj = RieszEnergySampling(*args, **kwargs)
     else:
         raise ValueError(
             f"Unknown sampling algorithm: {algorithm!r}. "
-            "Options: ['random', 'lhs', 'halton', 'sobol']."
+            "Options: ['random', 'lhs', 'halton', 'sobol', 'riesz']."
         )
 
     return _obj.sample(n_points, n_dim)